Move Feeback wrapper to higher level into InputFile
