+++ src/ui/molecules/InputFile/FilesSummary.tsx
@@ -45,7 +45,7 @@ const Container = styled.div`
   height: 25px;
   padding-bottom: 5px;
   align-items: center;
-  margin: 20px 0;
+  margin: 00px 0;
 `
 
 const Content = styled.span`
