+++ src/ui/molecules/InputFile/InputFile.tsx
@@ -1,10 +1,13 @@
-import React, { useState } from 'react'
+import React, { Fragment, useMemo, useState } from 'react'
 
 import { Controller } from 'react-hook-form'
+
+import { getFormFieldsErrors } from '../../../utils'
 
 import Form from '../../hocs/Form'
 import InputFileUploader from './Uploader'
 import FilesSummary from './FilesSummary'
+import InputFeedback from '../../hocs/InputFeedback'
 
 interface Props {
   component?: React.ReactNode,
@@ -14,53 +17,65 @@ interface Props {
   doNotShowFeedback?: boolean
   label?: string
   width?: string
+  accept: string
 }
 
 const defaultProps: Partial<Props> = {
   defaultValue: '',
   disabled: false,
-  width: '220px'
+  width: '220px',
+  accept: '.jpg, .jpeg, .png, .svg, .webp, application/pdf'
 }
 
 function InputFile({
-  component,
   name,
-  disabled,
-  defaultValue,
   label,
   width,
+  accept,
+  disabled,
+  component,
+  defaultValue,
   doNotShowFeedback
 }): JSX.Element {
   const [loadedFile, setLoadedFile] = useState<File | File[]>(null)
 
-  const { control } = Form.useForm()
+  const { control, formState: { errors } } = Form.useForm()
+
+  const fieldError: any = useMemo(() => getFormFieldsErrors(errors, name), [errors, name])
+  const hasError = useMemo(() => !!fieldError?.message, [fieldError])
+
+  const Container = useMemo(() => doNotShowFeedback ? Fragment : InputFeedback, [
+    doNotShowFeedback
+  ])
 
   return (
     <Controller
       name={name}
       control={control}
-      render={({ field: { onChange } }) => loadedFile ? (
-        <FilesSummary
-          clearFile={() => {
-            onChange?.(null)
-            setLoadedFile(null)
-          }}
-          file={loadedFile}
-        />
-      ) : (
-        <InputFileUploader
-          label={label}
-          name={name}
-          component={component}
-          disabled={disabled}
-          onChange={(file: File | File[]) => {
-            onChange?.(file)
-            setLoadedFile(file)
-          }}
-          width={width}
-          doNotShowFeedback={doNotShowFeedback}
-        />
-      )}
+      render={({ field: { onChange } }) =>
+        <Container label={label} errors={fieldError} hasError={hasError} name={name}>
+          {loadedFile ? (
+            <FilesSummary
+              file={loadedFile}
+              clearFile={() => {
+                onChange?.(null)
+                setLoadedFile(null)
+              }}
+            />
+          ) : (
+            <InputFileUploader
+              accept={accept}
+              width={width}
+              disabled={disabled}
+              component={component}
+              onChange={(file: File | File[]) => {
+                onChange?.(file)
+                setLoadedFile(file)
+              }}
+            />
+          )}
+        </Container>
+      }
       defaultValue={defaultValue}
     />
   )
